Add logic to users and answers

+++ db/migrate/20171108112439_create_users.rb
@@ -1,13 +1,15 @@
 class CreateUsers < ActiveRecord::Migration[5.1]
-  def change
-    drop_table :users
+    def change
+        if table_exists?("users")
+            drop_table :users
+        end
     
-    create_table :users do |t|
-      t.string :name
-      t.string :email
-      t.string :password_digest
+        create_table :users do |t|
+            t.string :name
+            t.string :email
+            t.string :password_digest
 
-      t.timestamps
+            t.timestamps
+        end
     end
-  end
 end
